change group_by semantics
#249 explained that other rx impls have a different group_by behavior.
This turned out to be a very large change. the lifetime coordination
required some thread safety changes. I tried a version of this change
that introduced a coordination parameter, but the only shared state was
the count, so this change makes that int atomic, even when no threads
are involved. not ideal but I decided it was the best tradeoff.