Work-around for VS2017RC error C2672 building doc (#357)
When building RxCPP with the latest VS2017 RC, the compiler emits error
C2672 on compiling a method.

There was some off-line discussion (Kirk Shoop was included) about the
legitimacy of the error. In either case, I'd anticipate that the error
could hit in VS2017RTW, so I'm offering this PR.

The change is very targeted -- a natural question is if this error could
arise in the methods around retry. For that and further discussion, I
have to point you to my colleagues participating in the off-line
discussion; for this particular patch all I could do was confirm the
latest RC still needs it to build.