Add CMake option to not build and run test and examples
It's super great that RxCpp has such a comprehensive set of tests and
examples, but when I include RxCpp in my own (CMake based) project,
I am not particularly interested in compiling them (which takes a
significant amount of time, especially compared to the rest of my
project) or running the tests with CTest.

CMake has a nice EXCLUDE_FROM_ALL option to its add_subdirectory
command which makes it build only things that you depend on explicitly,
but unfortunately it does not interact well with CTest. If I use
EXCLUDE_FROM_ALL on RxCpp, then ctest fails because it attempts to
run tests that were not built.

As far as I know there is no sane way to disable CTest tests once they
have been registered, and the idiomatic CMake approach seems to be to
have an option like this that optionally disables tests and similar
things. Flatbuffers do this, for example.

With this patch, I will be able to import RxCpp to my project like this:

set(RXCPP_BUILD_TESTS OFF)
add_subdirectory(rxcpp)