Include headers directory as SYSTEM path
When using the `rxcpp` target from a cmake subproject (using fetch_content or add_subdirectory) it is not possible to use `-Werror` or stricter warnings in general due to warnings from the rxcpp headers being reported. This change sets the `SYSTEM` flag only if the variable `RXCPP_INCLUDE_AS_SYSTEM_HEADERS` is set to true. In the end this passes the include directories to gcc using the `-i` flag instead of the `-I` flag and warnings from the rxcpp headers are not reported in projects including it. SYSTEM has to be set conditonally because that would also suppress valid warnings for the rxcpp tests and examples.